[2019-11-24]版本 v1.2
新特性:
(1)被依赖中的任务不能删除;
(2)手动执行任务时提供是否使用“任务依赖”选项;
(2)对Cron触发的任务调度、手动执行时启用任务依赖的任务调度，在运行失败后提供“重跑”按钮;
(3)执行器失去联系、关闭、重启后，相应执行器上的“进行中”的任务调度状态设置为“失败”;
(4)任务调度中心重启后，“待运行”的任务调度状态设置为“失败”;
(5)查看DAG里新增右键菜单“查看任务”、“复制任务名”

修复BUG:
(1)终止“待运行”的任务调度时仍重试;
(2)任务依赖偶尔不能精准监控上游任务当日最后一次运行的状态;
(3)任务重试时参数丢失;
(4)使用任务依赖的任务调度重试时任务依赖不生效;

优化:
(1)[from xxl-job]DB重连优化，修复DB宕机重连后任务调度停止的问题，重连后自动加入调度集群触发任务调度;

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/service/impl/XxlJobServiceImpl.java
@@ -310,6 +310,11 @@ public class XxlJobServiceImpl implements XxlJobService {
 			return ReturnT.SUCCESS;
 		}
 
+		int useJobs = xxlJobInfoDao.findCountByUpStreamJobList(id);
+		if(useJobs>0){
+			return new ReturnT<String>(ReturnT.FAIL_CODE, I18nUtil.getString("system_task_be_dependent_on")+","+I18nUtil.getString("system_opt_del")+I18nUtil.getString("system_fail"));
+		}
+
 		xxlJobInfoDao.delete(id);
 		xxlJobLogDao.delete(id);
 		xxlJobLogGlueDao.deleteByJobId(id);
